fixed spreading into different files instead of multiple ones
